added color parsing. TODO: camera movement, projections

+++ draw.c
@@ -6,7 +6,7 @@
 /*   By: sgoffaux <sgoffaux@student.s19.be>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/06/23 10:17:05 by sgoffaux          #+#    #+#             */
-/*   Updated: 2021/06/25 11:50:07 by sgoffaux         ###   ########.fr       */
+/*   Updated: 2021/06/25 14:57:56 by sgoffaux         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -46,7 +46,7 @@ void	ft_draw(t_map *map, t_fdf *env)
 	int	y;
 	int zoom;
 
-	zoom = ft_min((WIDTH / map->width / 2), (HEIGHT / map->height / 2));
+	zoom = ft_min((int)(WIDTH / map->width / 2.f), (int)(HEIGHT / map->height / 2.f));
 	ft_background(env);
 	y = 0;
 	while (y < map->height)
